requirements.txt
certifi==2020.12.5
chardet==4.0.0
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
openai==0.6.3
requests==2.25.1
smmap==4.0.0
tqdm==4.60.0
urllib3==1.26.4
